refactor(a8-web): centralize user-facing strings in a messages catalog
