+++ apps/a8-web/src/commands.ts
@@ -1,5 +1,6 @@
 import type { Emulator } from "./emulator.ts";
 import type { EmulatorHost } from "./host.ts";
+import { labels, type LabelKey } from "./messages.ts";
 
 export interface CommandContext {
 	/** The live machine — for the key-matrix and joystick commands. */
@@ -517,322 +518,6 @@ for (const key of Object.keys(commands) as Command[]) {
 
 /** Every bindable command name — the key-binding and palette surface. */
 export type Command = keyof typeof commands;
-
-/**
- * The flat label catalog. Today it is the English strings; it is shaped to
- * become a per-language translation table later (commands reference these keys,
- * not the strings). No interpolation yet — add it if/when a label needs it.
- */
-export const labels = {
-	POWER_CYCLE: "Power cycle (cold reset)",
-	PAUSE: "Pause emulation",
-	RESUME: "Resume emulation",
-	TOGGLE_PAUSE: "Pause or resume emulation",
-	TURBO_MODE_ENABLE: "Enable turbo mode (run unthrottled, muted)",
-	TURBO_MODE_DISABLE: "Disable turbo mode (return to real-time speed)",
-	TURBO_MODE_TOGGLE: "Toggle turbo mode",
-	AUDIO_MUTE: "Mute audio",
-	AUDIO_UNMUTE: "Unmute audio",
-	AUDIO_TOGGLE: "Toggle audio (enable, then mute/unmute)",
-	MENU_TOGGLE: "Toggle the menu",
-	BOOT_IMAGE: "Boot a disk, cartridge, or executable…",
-	ATTACH_D1: "Attach a disk to D1:…",
-	DETACH_D1: "Detach the disk from D1:",
-	ATTACH_CARTRIDGE: "Attach a cartridge… (reboots)",
-	DETACH_CARTRIDGE: "Detach the cartridge (reboots)",
-	DOWNLOAD_D1: "Download the D1: disk image…",
-	SET_TYPE_800: "Set machine type to Atari 800 (reboots)",
-	SET_TYPE_800XL: "Set machine type to Atari 800XL (reboots)",
-	SET_TYPE_130XE: "Set machine type to Atari 130XE (reboots)",
-	SET_TV_NTSC: "Set TV standard to NTSC (reboots)",
-	SET_TV_PAL: "Set TV standard to PAL (reboots)",
-	SET_TV_TOGGLE: "Toggle TV standard (NTSC/PAL) (reboots)",
-	BASIC_ENABLE: "Enable BASIC (reboots)",
-	BASIC_DISABLE: "Disable BASIC (reboots)",
-	BASIC_TOGGLE: "Toggle BASIC (reboots)",
-	PRESS_L: "Press L ($00)",
-	PRESS_J: "Press J ($01)",
-	PRESS_SEMICOLON: "Press Semicolon (';', $02)",
-	PRESS_F1: "Press F1 (Cursor Up, $03)",
-	PRESS_F2: "Press F2 (Cursor Down, $04)",
-	PRESS_K: "Press K ($05)",
-	PRESS_PLUS: "Press Plus ('+', $06)",
-	PRESS_ASTERISK: "Press Asterisk ('*', $07)",
-	PRESS_O: "Press O ($08)",
-	PRESS_CODE_09: "Press key $09 ($09)",
-	PRESS_P: "Press P ($0A)",
-	PRESS_U: "Press U ($0B)",
-	PRESS_RETURN: "Press Return ($0C)",
-	PRESS_I: "Press I ($0D)",
-	PRESS_MINUS: "Press Minus ('-', $0E)",
-	PRESS_EQUALS: "Press Equals ('=', $0F)",
-	PRESS_V: "Press V ($10)",
-	PRESS_HELP: "Press Help ($11)",
-	PRESS_C: "Press C ($12)",
-	PRESS_F3: "Press F3 (Cursor Left, $13)",
-	PRESS_F4: "Press F4 (Cursor Right, $14)",
-	PRESS_B: "Press B ($15)",
-	PRESS_X: "Press X ($16)",
-	PRESS_Z: "Press Z ($17)",
-	PRESS_4: "Press 4 ($18)",
-	PRESS_CODE_19: "Press key $19 ($19)",
-	PRESS_3: "Press 3 ($1A)",
-	PRESS_6: "Press 6 ($1B)",
-	PRESS_ESC: "Press Esc ($1C)",
-	PRESS_5: "Press 5 ($1D)",
-	PRESS_2: "Press 2 ($1E)",
-	PRESS_1: "Press 1 ($1F)",
-	PRESS_COMMA: "Press Comma (',', $20)",
-	PRESS_SPACE: "Press Space ($21)",
-	PRESS_PERIOD: "Press Period ('.', $22)",
-	PRESS_N: "Press N ($23)",
-	PRESS_CODE_24: "Press key $24 ($24)",
-	PRESS_M: "Press M ($25)",
-	PRESS_SLASH: "Press Slash ('/', $26)",
-	PRESS_INVERSE_VIDEO: "Press Inverse Video ($27)",
-	PRESS_R: "Press R ($28)",
-	PRESS_CODE_29: "Press key $29 ($29)",
-	PRESS_E: "Press E ($2A)",
-	PRESS_Y: "Press Y ($2B)",
-	PRESS_TAB: "Press Tab ($2C)",
-	PRESS_T: "Press T ($2D)",
-	PRESS_W: "Press W ($2E)",
-	PRESS_Q: "Press Q ($2F)",
-	PRESS_9: "Press 9 ($30)",
-	PRESS_CODE_31: "Press key $31 ($31)",
-	PRESS_0: "Press 0 ($32)",
-	PRESS_7: "Press 7 ($33)",
-	PRESS_BACKSPACE: "Press Backspace ($34)",
-	PRESS_8: "Press 8 ($35)",
-	PRESS_LESS_THAN: "Press Less Than ('<', $36)",
-	PRESS_GREATER_THAN: "Press Greater Than ('>', $37)",
-	PRESS_F: "Press F ($38)",
-	PRESS_H: "Press H ($39)",
-	PRESS_D: "Press D ($3A)",
-	PRESS_CODE_3B: "Press key $3B ($3B)",
-	PRESS_CAPS: "Press Caps ($3C)",
-	PRESS_G: "Press G ($3D)",
-	PRESS_S: "Press S ($3E)",
-	PRESS_A: "Press A ($3F)",
-	PRESS_SHIFT_L: "Press Shift+L ($40)",
-	PRESS_SHIFT_J: "Press Shift+J ($41)",
-	PRESS_SHIFT_SEMICOLON: "Press Shift+Semicolon (Colon, ':', $42)",
-	PRESS_SHIFT_F1: "Press Shift+F1 (Cursor to Upper Left Corner, $43)",
-	PRESS_SHIFT_F2: "Press Shift+F2 (Cursor to Lower Left Corner, $44)",
-	PRESS_SHIFT_K: "Press Shift+K ($45)",
-	PRESS_SHIFT_PLUS: "Press Shift+Plus (Backslash, '\\', $46)",
-	PRESS_SHIFT_ASTERISK: "Press Shift+Asterisk (Circumflex, '^', $47)",
-	PRESS_SHIFT_O: "Press Shift+O ($48)",
-	PRESS_SHIFT_CODE_09: "Press key $49 ($49)",
-	PRESS_SHIFT_P: "Press Shift+P ($4A)",
-	PRESS_SHIFT_U: "Press Shift+U ($4B)",
-	PRESS_SHIFT_RETURN: "Press Shift+Return ($4C)",
-	PRESS_SHIFT_I: "Press Shift+I ($4D)",
-	PRESS_SHIFT_MINUS: "Press Shift+Minus (Underscore, '_', $4E)",
-	PRESS_SHIFT_EQUALS: "Press Shift+Equals (Vertical Line, '|', $4F)",
-	PRESS_SHIFT_V: "Press Shift+V ($50)",
-	PRESS_SHIFT_HELP: "Press Shift+Help ($51)",
-	PRESS_SHIFT_C: "Press Shift+C ($52)",
-	PRESS_SHIFT_F3: "Press Shift+F3 (Cursor to Beginning of Line, $53)",
-	PRESS_SHIFT_F4: "Press Shift+F4 (Cursor to End of Line, $54)",
-	PRESS_SHIFT_B: "Press Shift+B ($55)",
-	PRESS_SHIFT_X: "Press Shift+X ($56)",
-	PRESS_SHIFT_Z: "Press Shift+Z ($57)",
-	PRESS_SHIFT_4: "Press Shift+4 (Dollar Sign, '$', $58)",
-	PRESS_SHIFT_CODE_19: "Press key $59 ($59)",
-	PRESS_SHIFT_3: "Press Shift+3 (Number Sign, '#', $5A)",
-	PRESS_SHIFT_6: "Press Shift+6 (Ampersand, '&', $5B)",
-	PRESS_SHIFT_ESC: "Press Shift+Esc ($5C)",
-	PRESS_SHIFT_5: "Press Shift+5 (Percent Sign, '%', $5D)",
-	PRESS_SHIFT_2: "Press Shift+2 (Quotation Mark, '\"', $5E)",
-	PRESS_SHIFT_1: "Press Shift+1 (Exclamation Mark, '!', $5F)",
-	PRESS_SHIFT_COMMA: "Press Shift+Comma (Left Square Bracket, '[', $60)",
-	PRESS_SHIFT_SPACE: "Press Shift+Space ($61)",
-	PRESS_SHIFT_PERIOD: "Press Shift+Period (Right Square Bracket, ']', $62)",
-	PRESS_SHIFT_N: "Press Shift+N ($63)",
-	PRESS_SHIFT_CODE_24: "Press key $64 ($64)",
-	PRESS_SHIFT_M: "Press Shift+M ($65)",
-	PRESS_SHIFT_SLASH: "Press Shift+Slash (Question Mark, '?', $66)",
-	PRESS_SHIFT_INVERSE_VIDEO: "Press Shift+Inverse Video ($67)",
-	PRESS_SHIFT_R: "Press Shift+R ($68)",
-	PRESS_SHIFT_CODE_29: "Press key $69 ($69)",
-	PRESS_SHIFT_E: "Press Shift+E ($6A)",
-	PRESS_SHIFT_Y: "Press Shift+Y ($6B)",
-	PRESS_SHIFT_TAB: "Press Shift+Tab (Set Tab Stop, $6C)",
-	PRESS_SHIFT_T: "Press Shift+T ($6D)",
-	PRESS_SHIFT_W: "Press Shift+W ($6E)",
-	PRESS_SHIFT_Q: "Press Shift+Q ($6F)",
-	PRESS_SHIFT_9: "Press Shift+9 (Left Parenthesis, '(', $70)",
-	PRESS_SHIFT_CODE_31: "Press key $71 ($71)",
-	PRESS_SHIFT_0: "Press Shift+0 (Right Parenthesis, ')', $72)",
-	PRESS_SHIFT_7: "Press Shift+7 (Apostrophe, ''', $73)",
-	PRESS_SHIFT_BACKSPACE: "Press Shift+Backspace (Delete Line, $74)",
-	PRESS_SHIFT_8: "Press Shift+8 (Commercial At, '@', $75)",
-	PRESS_SHIFT_LESS_THAN: "Press Shift+Less Than (Clear Screen, $76)",
-	PRESS_SHIFT_GREATER_THAN: "Press Shift+Greater Than (Insert Line, $77)",
-	PRESS_SHIFT_F: "Press Shift+F ($78)",
-	PRESS_SHIFT_H: "Press Shift+H ($79)",
-	PRESS_SHIFT_D: "Press Shift+D ($7A)",
-	PRESS_SHIFT_CODE_3B: "Press key $7B ($7B)",
-	PRESS_SHIFT_CAPS: "Press Shift+Caps ($7C)",
-	PRESS_SHIFT_G: "Press Shift+G ($7D)",
-	PRESS_SHIFT_S: "Press Shift+S ($7E)",
-	PRESS_SHIFT_A: "Press Shift+A ($7F)",
-	PRESS_CONTROL_L: "Press Control+L (Upper Left Quadrant, '▘', $80)",
-	PRESS_CONTROL_J: "Press Control+J (Lower Left Triangle, '◣', $81)",
-	PRESS_CONTROL_SEMICOLON: "Press Control+Semicolon (Spade, '♠', $82)",
-	PRESS_CONTROL_F1: "Press Control+F1 (Keyboard Enable/Disable, $83)",
-	PRESS_CONTROL_F2: "Press Control+F2 (Screen DMA Enable/Disable, $84)",
-	PRESS_CONTROL_K: "Press Control+K (Upper Right Quadrant, '▝', $85)",
-	PRESS_CONTROL_PLUS: "Press Control+Plus (Cursor Left, $86)",
-	PRESS_CONTROL_ASTERISK: "Press Control+Asterisk (Cursor Right, $87)",
-	PRESS_CONTROL_O: "Press Control+O (Lower Left Quadrant, '▖', $88)",
-	PRESS_CONTROL_CODE_89: "Press key $89 ($89)",
-	PRESS_CONTROL_P: "Press Control+P (Club, '♣', $8A)",
-	PRESS_CONTROL_U: "Press Control+U (Lower Half Block, '▄', $8B)",
-	PRESS_CONTROL_RETURN: "Press Control+Return ($8C)",
-	PRESS_CONTROL_I: "Press Control+I (Lower Right Quadrant, '▗', $8D)",
-	PRESS_CONTROL_MINUS: "Press Control+Minus (Cursor Up, $8E)",
-	PRESS_CONTROL_EQUALS: "Press Control+Equals (Cursor Down, $8F)",
-	PRESS_CONTROL_V: "Press Control+V (Left Vertical Bar, '▏', $90)",
-	PRESS_CONTROL_HELP: "Press Control+Help ($91)",
-	PRESS_CONTROL_C: "Press Control+C (Box Up and Left, '┘', $92)",
-	PRESS_CONTROL_F3: "Press Control+F3 (Toggle Key Click, $93)",
-	PRESS_CONTROL_F4: "Press Control+F4 (Toggle Character Set, $94)",
-	PRESS_CONTROL_B: "Press Control+B (Right Vertical Bar, '▕', $95)",
-	PRESS_CONTROL_X: "Press Control+X (Box Up and Horizontal, '┴', $96)",
-	PRESS_CONTROL_Z: "Press Control+Z (Box Up and Right, '└', $97)",
-	PRESS_CONTROL_4: "Press Control+4 ($98)",
-	PRESS_CONTROL_CODE_99: "Press key $99 ($99)",
-	PRESS_CONTROL_3: "Press Control+3 (EOF, $9A)",
-	PRESS_CONTROL_6: "Press Control+6 ($9B)",
-	PRESS_CONTROL_ESC: "Press Control+Esc ($9C)",
-	PRESS_CONTROL_5: "Press Control+5 ($9D)",
-	PRESS_CONTROL_2: "Press Control+2 (Buzzer, $9E)",
-	PRESS_CONTROL_1: "Press Control+1 (Pause/Resume Screen Output, $9F)",
-	PRESS_CONTROL_COMMA: "Press Control+Comma (Heart, '♥', $A0)",
-	PRESS_CONTROL_SPACE: "Press Control+Space ($A1)",
-	PRESS_CONTROL_PERIOD: "Press Control+Period (Diamond, '♦', $A2)",
-	PRESS_CONTROL_N: "Press Control+N (Lower Horizontal Bar, '▁', $A3)",
-	PRESS_CONTROL_CODE_A4: "Press key $A4 ($A4)",
-	PRESS_CONTROL_M: "Press Control+M (Upper Horizontal Bar, '▔', $A5)",
-	PRESS_CONTROL_SLASH: "Press Control+Slash ($A6)",
-	PRESS_CONTROL_INVERSE_VIDEO:
-		"Press Control+Inverse Video (Control Lock, $A7)",
-	PRESS_CONTROL_R: "Press Control+R (Horizontal Bar, '─', $A8)",
-	PRESS_CONTROL_CODE_A9: "Press key $A9 ($A9)",
-	PRESS_CONTROL_E: "Press Control+E (Box Down and Left, '┐', $AA)",
-	PRESS_CONTROL_Y: "Press Control+Y (Left Half Block, '▌', $AB)",
-	PRESS_CONTROL_TAB: "Press Control+Tab (Clear Tab Stop, $AC)",
-	PRESS_CONTROL_T: "Press Control+T (Bullet, '•', $AD)",
-	PRESS_CONTROL_W: "Press Control+W (Box Down and Horizontal, '┬', $AE)",
-	PRESS_CONTROL_Q: "Press Control+Q (Box Down and Right, '┌', $AF)",
-	PRESS_CONTROL_9: "Press Control+9 ($B0)",
-	PRESS_CONTROL_CODE_B1: "Press key $B1 ($B1)",
-	PRESS_CONTROL_0: "Press Control+0 ($B2)",
-	PRESS_CONTROL_7: "Press Control+7 ($B3)",
-	PRESS_CONTROL_BACKSPACE: "Press Control+Backspace (Delete, $B4)",
-	PRESS_CONTROL_8: "Press Control+8 ($B5)",
-	PRESS_CONTROL_LESS_THAN: "Press Control+Less Than (Clear Screen, $B6)",
-	PRESS_CONTROL_GREATER_THAN:
-		"Press Control+Greater Than (Insert Character, $B7)",
-	PRESS_CONTROL_F: "Press Control+F (Upper Right to Lower Left, '╱', $B8)",
-	PRESS_CONTROL_H: "Press Control+H (Lower Right Triangle, '◢', $B9)",
-	PRESS_CONTROL_D: "Press Control+D (Box Vertical and Left, '┤', $BA)",
-	PRESS_CONTROL_CODE_BB: "Press key $BB ($BB)",
-	PRESS_CONTROL_CAPS: "Press Control+Caps ($BC)",
-	PRESS_CONTROL_G: "Press Control+G (Upper Left to Lower Right, '╲', $BD)",
-	PRESS_CONTROL_S: "Press Control+S (Box Vertical and Horizontal, '┼', $BE)",
-	PRESS_CONTROL_A: "Press Control+A (Box Vertical and Right, '├', $BF)",
-	PRESS_CONTROL_SHIFT_L: "Press Control+Shift+L ($C0)",
-	PRESS_CONTROL_SHIFT_J: "Press Control+Shift+J ($C1)",
-	PRESS_CONTROL_SHIFT_SEMICOLON: "Press Control+Shift+Semicolon ($C2)",
-	PRESS_CONTROL_SHIFT_F1: "Press Control+Shift+F1 ($C3)",
-	PRESS_CONTROL_SHIFT_F2: "Press Control+Shift+F2 ($C4)",
-	PRESS_CONTROL_SHIFT_K: "Press Control+Shift+K ($C5)",
-	PRESS_CONTROL_SHIFT_PLUS: "Press Control+Shift+Plus ($C6)",
-	PRESS_CONTROL_SHIFT_ASTERISK: "Press Control+Shift+Asterisk ($C7)",
-	PRESS_CONTROL_SHIFT_O: "Press Control+Shift+O ($C8)",
-	PRESS_CONTROL_SHIFT_CODE_09: "Press key $C9 ($C9)",
-	PRESS_CONTROL_SHIFT_P: "Press Control+Shift+P ($CA)",
-	PRESS_CONTROL_SHIFT_U: "Press Control+Shift+U ($CB)",
-	PRESS_CONTROL_SHIFT_RETURN: "Press Control+Shift+Return ($CC)",
-	PRESS_CONTROL_SHIFT_I: "Press Control+Shift+I ($CD)",
-	PRESS_CONTROL_SHIFT_MINUS: "Press Control+Shift+Minus ($CE)",
-	PRESS_CONTROL_SHIFT_EQUALS: "Press Control+Shift+Equals ($CF)",
-	PRESS_CONTROL_SHIFT_V: "Press Control+Shift+V ($D0)",
-	PRESS_CONTROL_SHIFT_HELP: "Press Control+Shift+Help ($D1)",
-	PRESS_CONTROL_SHIFT_C: "Press Control+Shift+C ($D2)",
-	PRESS_CONTROL_SHIFT_F3: "Press Control+Shift+F3 ($D3)",
-	PRESS_CONTROL_SHIFT_F4: "Press Control+Shift+F4 ($D4)",
-	PRESS_CONTROL_SHIFT_B: "Press Control+Shift+B ($D5)",
-	PRESS_CONTROL_SHIFT_X: "Press Control+Shift+X ($D6)",
-	PRESS_CONTROL_SHIFT_Z: "Press Control+Shift+Z ($D7)",
-	PRESS_CONTROL_SHIFT_4: "Press Control+Shift+4 ($D8)",
-	PRESS_CONTROL_SHIFT_CODE_19: "Press key $D9 ($D9)",
-	PRESS_CONTROL_SHIFT_3: "Press Control+Shift+3 ($DA)",
-	PRESS_CONTROL_SHIFT_6: "Press Control+Shift+6 ($DB)",
-	PRESS_CONTROL_SHIFT_ESC: "Press Control+Shift+Esc ($DC)",
-	PRESS_CONTROL_SHIFT_5: "Press Control+Shift+5 ($DD)",
-	PRESS_CONTROL_SHIFT_2: "Press Control+Shift+2 ($DE)",
-	PRESS_CONTROL_SHIFT_1: "Press Control+Shift+1 ($DF)",
-	PRESS_CONTROL_SHIFT_COMMA: "Press Control+Shift+Comma ($E0)",
-	PRESS_CONTROL_SHIFT_SPACE: "Press Control+Shift+Space ($E1)",
-	PRESS_CONTROL_SHIFT_PERIOD: "Press Control+Shift+Period ($E2)",
-	PRESS_CONTROL_SHIFT_N: "Press Control+Shift+N ($E3)",
-	PRESS_CONTROL_SHIFT_CODE_24: "Press key $E4 ($E4)",
-	PRESS_CONTROL_SHIFT_M: "Press Control+Shift+M ($E5)",
-	PRESS_CONTROL_SHIFT_SLASH: "Press Control+Shift+Slash ($E6)",
-	PRESS_CONTROL_SHIFT_INVERSE_VIDEO: "Press Control+Shift+Inverse Video ($E7)",
-	PRESS_CONTROL_SHIFT_R: "Press Control+Shift+R ($E8)",
-	PRESS_CONTROL_SHIFT_CODE_29: "Press key $E9 ($E9)",
-	PRESS_CONTROL_SHIFT_E: "Press Control+Shift+E ($EA)",
-	PRESS_CONTROL_SHIFT_Y: "Press Control+Shift+Y ($EB)",
-	PRESS_CONTROL_SHIFT_TAB: "Press Control+Shift+Tab ($EC)",
-	PRESS_CONTROL_SHIFT_T: "Press Control+Shift+T ($ED)",
-	PRESS_CONTROL_SHIFT_W: "Press Control+Shift+W ($EE)",
-	PRESS_CONTROL_SHIFT_Q: "Press Control+Shift+Q ($EF)",
-	PRESS_CONTROL_SHIFT_9: "Press Control+Shift+9 ($F0)",
-	PRESS_CONTROL_SHIFT_CODE_31: "Press key $F1 ($F1)",
-	PRESS_CONTROL_SHIFT_0: "Press Control+Shift+0 ($F2)",
-	PRESS_CONTROL_SHIFT_7: "Press Control+Shift+7 ($F3)",
-	PRESS_CONTROL_SHIFT_BACKSPACE: "Press Control+Shift+Backspace ($F4)",
-	PRESS_CONTROL_SHIFT_8: "Press Control+Shift+8 ($F5)",
-	PRESS_CONTROL_SHIFT_LESS_THAN: "Press Control+Shift+Less Than ($F6)",
-	PRESS_CONTROL_SHIFT_GREATER_THAN: "Press Control+Shift+Greater Than ($F7)",
-	PRESS_CONTROL_SHIFT_F: "Press Control+Shift+F ($F8)",
-	PRESS_CONTROL_SHIFT_H: "Press Control+Shift+H ($F9)",
-	PRESS_CONTROL_SHIFT_D: "Press Control+Shift+D ($FA)",
-	PRESS_CONTROL_SHIFT_CODE_3B: "Press key $FB ($FB)",
-	PRESS_CONTROL_SHIFT_CAPS: "Press Control+Shift+Caps ($FC)",
-	PRESS_CONTROL_SHIFT_G: "Press Control+Shift+G ($FD)",
-	PRESS_CONTROL_SHIFT_S: "Press Control+Shift+S ($FE)",
-	PRESS_CONTROL_SHIFT_A: "Press Control+Shift+A ($FF)",
-	RELEASE_POKEY_KEY: "Release POKEY key",
-	PRESS_SHIFT: "Press Shift",
-	RELEASE_SHIFT: "Release Shift",
-	PRESS_RESET: "Press Reset",
-	RELEASE_RESET: "Release Reset",
-	PRESS_OPTION: "Press Option",
-	RELEASE_OPTION: "Release Option",
-	PRESS_SELECT: "Press Select",
-	RELEASE_SELECT: "Release Select",
-	PRESS_START: "Press Start",
-	RELEASE_START: "Release Start",
-	PRESS_BREAK: "Press Break",
-	PRESS_JOY0_UP: "Push joystick 0 up",
-	RELEASE_JOY0_UP: "Release joystick 0 up",
-	PRESS_JOY0_DOWN: "Push joystick 0 down",
-	RELEASE_JOY0_DOWN: "Release joystick 0 down",
-	PRESS_JOY0_LEFT: "Push joystick 0 left",
-	RELEASE_JOY0_LEFT: "Release joystick 0 left",
-	PRESS_JOY0_RIGHT: "Push joystick 0 right",
-	RELEASE_JOY0_RIGHT: "Release joystick 0 right",
-	PRESS_JOY0_TRIGGER: "Press joystick 0 trigger",
-	RELEASE_JOY0_TRIGGER: "Release joystick 0 trigger",
-} satisfies Record<string, string>;
-
-export type LabelKey = keyof typeof labels;
 
 /** The display label for a command (its current-language string). */
 export function labelOf(command: Command): string {
